move keys to a sub-table per user
